build: move pretty-quick step to pkg
