+++ packages/scripts/src/steps/index.ts
@@ -3,4 +3,3 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 export * from './eslint'
-export * from './pretty-quick'
