+++ packages/steps/pretty-quick/.npmignore
@@ -1,0 +1,1 @@
+*.tsbuildinfo
