+++ packages/steps/pretty-quick/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../../../tsconfig.json",
+	"compilerOptions": {
+		"outDir": "lib"
+	},
+	"include": ["src/**/*.ts"]
+}
